requirements.txt
# Quick mafs
numpy

# For saving/loading models
pickle

# Visualize (use later)
matplotlib

# Specifically for the digits dataset (preproc)
Pillow

